refactor: split server tests into separate singleton module
test_exiftool_server_client.py: 24 tests (server, client, port file,
integration, spawn)
test_exiftool_server_singleton.py: 7 tests (lock acquire, idempotent,
higher-PID conflict, lock cleanup, PID takeover, stress concurrent)

Clean up unused imports in both files.
