+++ tests/test_exiftool_server_singleton.py
@@ -1,0 +1,247 @@
+# -*- coding: utf-8 -*-
+"""
+Tests for ExifToolServer singleton lock and PID election.
+"""
+
+import json
+import os
+import socket
+import sys
+import tempfile
+import time
+import unittest
+
+import exiftool
+from exiftool.exceptions import ExifToolServerError
+
+
+@unittest.skipIf(
+	os.environ.get("SKIP_SINGLETON_TESTS"),
+	"Skipping singleton tests that use subprocess",
+)
+class TestExifToolServerSingleton(unittest.TestCase):
+	"""Test singleton lock enforcement on ExifToolServer."""
+
+	@classmethod
+	def setUpClass(cls):
+		cls.port_file = os.path.join(
+			tempfile.gettempdir(), "pyexiftool-test-singleton.json")
+		cls.lock_file = cls.port_file + ".lock"
+		for p in (cls.port_file, cls.lock_file):
+			try:
+				os.unlink(p)
+			except OSError:
+				pass
+		cls._own_pid = os.getpid()
+
+	@classmethod
+	def tearDownClass(cls):
+		for p in (cls.port_file, cls.lock_file):
+			try:
+				os.unlink(p)
+			except OSError:
+				pass
+
+	def setUp(self):
+		self.server = exiftool.ExifToolServer(
+			port_file=self.port_file,
+			singleton=True,
+			no_exiftool=True,
+		)
+
+	def tearDown(self):
+		self.server.stop()
+
+	def test_singleton_lock_acquired(self):
+		"""Start with singleton=True should create and hold the lock file."""
+		port = self.server.start()
+		self.assertGreater(port, 0)
+		self.assertTrue(os.path.exists(self.lock_file))
+
+	def test_singleton_idempotent(self):
+		"""Starting an already-running singleton server is idempotent."""
+		p1 = self.server.start()
+		p2 = self.server.start()
+		self.assertEqual(p1, p2)
+
+	def test_singleton_second_fails(self):
+		"""A second server with a higher PID must raise when the lock is held."""
+		self.server.start()
+		import unittest.mock as mock
+		real_pid = os.getpid()
+		fake_higher_pid = real_pid + 1000000
+		with mock.patch("exiftool.server.os.getpid", return_value=fake_higher_pid):
+			server2 = exiftool.ExifToolServer(
+				port_file=self.port_file,
+				singleton=True,
+				no_exiftool=True,
+			)
+			with self.assertRaises(ExifToolServerError) as cm:
+				server2.start()
+			self.assertIn("already running", str(cm.exception).lower())
+
+	def test_singleton_lock_cleaned_up(self):
+		"""After stop() the lock file must be released (re-acquirable)."""
+		self.server.start()
+		self.server.stop()
+		server2 = exiftool.ExifToolServer(
+			port_file=self.port_file,
+			singleton=True,
+			no_exiftool=True,
+		)
+		try:
+			port2 = server2.start()
+			self.assertGreater(port2, 0)
+		finally:
+			server2.stop()
+
+	def test_pid_takeover(self):
+		"""A server with a lower PID takes over from a higher-PID server."""
+		import subprocess
+
+		sub_code = (
+			"import sys; sys.path.insert(0, %r); import exiftool;"
+			"srv = exiftool.ExifToolServer("
+			"port_file=%r, singleton=True, no_exiftool=True,"
+			"idle_timeout=30); "
+			"srv.start()\n"
+			"while srv.running:\n"
+			"    import time; time.sleep(0.5)"
+		) % (
+			os.path.join(os.path.dirname(__file__), ".."),
+			self.port_file,
+		)
+		proc = subprocess.Popen(
+			[sys.executable, "-c", sub_code],
+			stdout=subprocess.DEVNULL,
+			stderr=subprocess.DEVNULL,
+		)
+		try:
+			deadline = time.monotonic() + 10.0
+			while time.monotonic() < deadline:
+				try:
+					with open(self.port_file) as f:
+						data = json.load(f)
+					port = data["port"]
+					with socket.create_connection(
+						("127.0.0.1", port), timeout=1.0) as s:
+						s.sendall(json.dumps(
+							{"id": 1, "method": "ping",
+							 "params": {}}).encode() + b"\n")
+						resp = s.makefile("r", encoding="utf-8").readline()
+					if resp and '"pong"' in resp:
+						break
+				except (OSError, KeyError, socket.timeout, ConnectionError):
+					pass
+				time.sleep(0.05)
+			self.assertIsNone(proc.poll(), "subprocess server died prematurely")
+
+			takeover = exiftool.ExifToolServer(
+				port_file=self.port_file,
+				singleton=True,
+				no_exiftool=True,
+			)
+			try:
+				port = takeover.start()
+				self.assertGreater(port, 0)
+				proc.wait(timeout=10.0)
+			except subprocess.TimeoutExpired:
+				self.fail("subprocess should have been terminated by takeover")
+			finally:
+				takeover.stop()
+		finally:
+			if proc.poll() is None:
+				proc.terminate()
+				proc.wait()
+
+	def test_singleton_stress_concurrent(self):
+		"""N concurrent processes with singleton=True: exactly one survives."""
+		N = 10
+		root = os.path.join(os.path.dirname(__file__), "..")
+
+		sub_code = (
+			"import sys, json, os\n"
+			"sys.path.insert(0, %r)\n"
+			"import exiftool\n"
+			"port_file = %r\n"
+			"result = {'pid': os.getpid(), 'status': 'unknown'}\n"
+			"try:\n"
+			"    srv = exiftool.ExifToolServer(\n"
+			"        port_file=port_file, singleton=True,\n"
+			"        no_exiftool=True, idle_timeout=30)\n"
+			"    port = srv.start()\n"
+			"    result['status'] = 'running'\n"
+			"    result['port'] = port\n"
+			"    while srv.running:\n"
+			"        import time; time.sleep(0.5)\n"
+			"    result['status'] = 'stopped'\n"
+			"except Exception as e:\n"
+			"    result['status'] = 'failed'\n"
+			"    result['error'] = type(e).__name__\n"
+			"print(json.dumps(result))\n"
+		) % (root, self.port_file)
+
+		import subprocess
+		procs = []
+		for i in range(N):
+			p = subprocess.Popen(
+				[sys.executable, "-c", sub_code],
+				stdout=subprocess.PIPE, stderr=subprocess.DEVNULL,
+			)
+			procs.append(p)
+			time.sleep(0.01)
+
+		deadline = time.monotonic() + 5.0
+		survivors = []
+		exited_results = []
+
+		for p in procs:
+			remaining = deadline - time.monotonic()
+			if remaining <= 0:
+				remaining = 0.001
+			try:
+				out, _ = p.communicate(timeout=remaining)
+				exited_results.append(json.loads(out.decode()))
+			except subprocess.TimeoutExpired:
+				survivors.append((p, p.pid))
+
+		self.assertGreater(len(survivors), 0,
+			f"Expected at least 1 survivor, got {len(survivors)}")
+
+		with open(self.port_file) as f:
+			port_data = json.load(f)
+		port = port_data["port"]
+		with socket.create_connection(("127.0.0.1", port), timeout=5) as s:
+			s.sendall(json.dumps(
+				{"id": 1, "method": "ping",
+				 "params": {}}).encode() + b"\n")
+			resp = s.makefile("r", encoding="utf-8").readline()
+		self.assertIn("pong", resp or "")
+
+		failed = [r for r in exited_results if r['status'] == 'failed']
+		self.assertGreater(len(failed), 0,
+			f"No failures — expected at least one higher-PID contention loss "
+			f"({len(exited_results)} exited, {len(survivors)} running)")
+
+		real_survivor = None
+		for proc, pid in survivors:
+			if proc.poll() is None:
+				if real_survivor is None:
+					real_survivor = (proc, pid)
+				else:
+					proc.terminate()
+		self.assertIsNotNone(real_survivor, "No surviving process found")
+
+		surv_proc, surv_pid = real_survivor
+		with socket.create_connection(("127.0.0.1", port), timeout=5) as s:
+			s.sendall(json.dumps(
+				{"id": 1, "method": "shutdown",
+				 "params": {}}).encode() + b"\n")
+
+		out, _ = surv_proc.communicate(timeout=10.0)
+		surv_result = json.loads(out.decode())
+		self.assertEqual(surv_result['status'], 'stopped')
+
+
+if __name__ == '__main__':
+	unittest.main()
